+++ detectors/text/__init__.py
@@ -1,0 +1,1 @@
+"""Text detectors for package-internal manuscript/source screening."""